Move task to new location
